config: allow fractional scheduler requeue sleep values
closes https://github.com/reanahub/reana-server/issues/407

+++ reana_server/config.py
@@ -180,5 +180,5 @@ ADMIN_EMAIL = os.getenv("REANA_EMAIL_SENDER", "CHANGE_ME")
 
 # Workflow scheduler
 # ==================
-REANA_SCHEDULER_REQUEUE_SLEEP = int(os.getenv("REANA_SCHEDULER_REQUEUE_SLEEP", "15"))
+REANA_SCHEDULER_REQUEUE_SLEEP = float(os.getenv("REANA_SCHEDULER_REQUEUE_SLEEP", "15"))
 """How many seconds to wait between consuming workflows."""
